+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -1,5 +1,6 @@
 package org.apache.coyote.http11.request;
 
+import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.HttpHeaders;
 import org.apache.coyote.http11.HttpMethod;
 import java.io.BufferedReader;
