+++ tomcat/src/main/java/org/apache/coyote/http11/response/HttpResponse.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.response;
 
-import org.apache.coyote.HttpStatus;
+import org.apache.coyote.http11.HttpStatus;
 import org.apache.coyote.http11.HttpHeaders;
 import org.apache.coyote.http11.request.HttpRequest;
 import java.io.IOException;
